# api/openapi.yaml
openapi: 3.1.0
info:
  title: Go SNMP OLT ZTE C320 Monitoring API
  description: |
    REST API for monitoring ZTE C320 and C300 OLT (Optical Line Terminal) via SNMP.
    Provides endpoints to query ONU status, optical power levels, serial numbers,
    and manage cache for board/PON combinations.

    **Multi-OLT:** when the instance is configured with multiple OLTs (the `OLTS`
    env var), every `/api/v1/board/...` path below is also available under
    `/api/v1/olt/{olt_id}/board/...`, where `{olt_id}` selects the target OLT.
    The configured default OLT continues to answer the bare `/api/v1/board/...`
    paths shown here. `board_id` is the physical GPON slot and `pon_id` is
    validated against that slot's card (GTGO=8 or GTGH=16 PON ports).

    **Breaking change in 3.0.0:** the JSON response format was aligned to
    the ISP adapter standard:
      - success responses now use `status: "success"` (was `"OK"`)
      - error responses use top-level `error_code` (was nested `error.type`)
      - error body carries `data` (was nested `error.message`) and `request_id`

    API contract version is `v1` (endpoint prefix `/api/v1`). The binary
    version and contract version are exposed via `X-App-Version` and
    `X-API-Version` response headers on every request.
  version: 3.2.0
  contact:
    name: API Support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/board/{board_id}/pon/{pon_id}:
    get:
      summary: Get all ONUs for a board/PON combination
      description: Retrieves a list of all ONUs connected to the specified board and PON port.
      operationId: getOnusByBoardPon
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: List of ONUs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ONUInfoPerBoard"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/cache/clear:
    delete:
      summary: Clear cache for a board/PON combination
      description: Removes the cached ONU data for the specified board and PON port. The next request will fetch fresh data from SNMP.
      operationId: clearCacheByBoardPon
      tags:
        - Cache
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Cache deleted successfully
                          board_id:
                            type: integer
                            example: 1
                          pon_id:
                            type: integer
                            example: 1
                        required:
                          - message
                          - board_id
                          - pon_id
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu/{onu_id}:
    get:
      summary: Get specific ONU detail
      description: Retrieves detailed information for a specific ONU including optical power levels, uptime, and offline history.
      operationId: getOnuDetail
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - $ref: "#/components/parameters/OnuId"
      responses:
        "200":
          description: ONU detail retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ONUCustomerInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id/empty:
    get:
      summary: Get available/empty ONU IDs
      description: Returns a list of ONU IDs that are not currently assigned on the specified board and PON port.
      operationId: getEmptyOnuIds
      tags:
        - ONU ID
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: List of empty ONU IDs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OnuID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id_sn:
    get:
      summary: Get ONU IDs with serial numbers
      description: Returns a list of ONU IDs along with their corresponding serial numbers for the specified board and PON port.
      operationId: getOnuIdSerialNumbers
      tags:
        - ONU ID
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - $ref: "#/components/parameters/NoCache"
      responses:
        "200":
          description: ONU IDs with serial numbers retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OnuSerialNumber"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id/update:
    post:
      summary: Force refresh empty ONU ID cache
      description: Triggers a cache refresh for empty ONU IDs on the specified board and PON port.
      operationId: updateEmptyOnuIdCache
      tags:
        - Cache
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: Cache updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: string
                        example: Success Update Empty ONU_ID
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/paginate/board/{board_id}/pon/{pon_id}:
    get:
      summary: Get paginated ONU list
      description: Retrieves a paginated list of ONUs for the specified board and PON port with metadata about total rows and page count.
      operationId: getOnusPaginated
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        "200":
          description: Paginated list of ONUs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ONUInfoPerBoard"
                      meta:
                        $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # ───────────────────────────────────────────────────────────────────────────
  # Multi-OLT variants. Identical semantics to the bare /api/v1/board/... paths
  # above, but {olt_id} selects which OLT in the OLTS/OLTS_FILE registry to
  # target. The configured DEFAULT_OLT also answers the bare paths above.
  # ───────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key for /api/v1 routes. With a single `API_KEY` the key only gates
        access. With the per-tenant `API_USERS` registry, the key also identifies
        the caller (user_id + role): a user sees only the OLTs whose `user_id`
        matches theirs — requesting another tenant's OLT returns 404 (not 403, to
        avoid enumeration) — while an `admin` role sees all OLTs. A missing or
        unknown key returns 401 when authentication is enabled.
  parameters:
...
    BoardId:
      name: board_id
      in: path
      required: true
      description: >-
        Physical GPON slot on the OLT (board). Which slots are valid depends on
        the deployment's OLT_BOARDS setting: a ZTE C320 typically uses slots 1-2,
        while a C300 chassis holds its GPON cards in higher slots (e.g. 3 and 5).
        Requests for a slot not configured via OLT_BOARDS return 400.
      schema:
        type: integer
        minimum: 1
        maximum: 30
        example: 1
    PonId:
      name: pon_id
      in: path
      required: true
      description: PON port identifier on the board
      schema:
        type: integer
        minimum: 1
        maximum: 16
        example: 1
    OnuId:
      name: onu_id
      in: path
      required: true
      description: ONU identifier on the PON port
      schema:
        type: integer
        minimum: 1
        maximum: 128
        example: 1
    NoCache:
      name: nocache
      in: query
      required: false
      description: >-
        When `true`, force a fresh SNMP read, bypassing the serial-list Redis
        cache (`board_<b>_pon_<p>_serial_list`) and refreshing it with the result.
        Used by write-olt-zte's pre-write existence checks so a delete/replace
        issued right after a provision sees the OLT's real state instead of a
        stale cached snapshot. Any value other than `true` keeps the default
        cached behaviour.
      schema:
        type: boolean
        default: false
        example: true
  schemas:
    WebResponse:
      type: object
      description: Standard success response wrapper
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Fixed string "success" for 2xx responses (ISP adapter standard)
          example: success
        data: {}
      required:
        - code
        - status
        - data
    ErrorResponse:
      type: object
      description: |
        Standard error response wrapper aligned to the `isp-adapter-standard`
        wiki page. `error_code` is a machine-readable identifier, `data`
        carries either the error message (string) or a structured object with
        `message` + `details`, and `request_id` echoes the request correlation
        ID so operators can correlate across services.
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        status:
          type: string
          description: HTTP reason phrase (e.g. "Bad Request", "Not Found")
          example: Bad Request
        error_code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - NOT_FOUND
            - UNAUTHORIZED
            - SNMP_ERROR
            - REDIS_ERROR
            - CONFIG_ERROR
            - INTERNAL_ERROR
          example: VALIDATION_ERROR
        data:
          description: |
            Either the human-readable error message (string) or an object
            containing `message` + `details` when structured context is
            available. Callers should handle both shapes.
          oneOf:
            - type: string
              example: board_id is not a configured GPON slot
            - type: object
              properties:
                message:
                  type: string
                  example: board_id is not a configured GPON slot
                details:
                  type: object
                  additionalProperties: true
                  example:
                    received: "99"
        request_id:
          type: string
          description: Correlation ID (same value as X-Request-ID response header)
          example: d7dhp4pciod1cuth4i80
      required:
        - code
        - status
        - error_code
        - data
    Meta:
      type: object
      description: Pagination metadata
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 10
        page_count:
          type: integer
          description: Total number of pages
          example: 7
        total_rows:
          type: integer
          description: Total number of records
          example: 69
      required:
        - page
        - limit
        - page_count
        - total_rows
    ONUInfoPerBoard:
      type: object
      description: Summary information for an ONU on a board/PON combination
      properties:
        board:
          type: integer
          description: Board number
          example: 1
        pon:
          type: integer
          description: PON port number
          example: 1
        onu_id:
          type: integer
          description: ONU identifier
          example: 1
        name:
          type: string
          description: Customer or ONU name
          example: Customer-001
        onu_type:
          type: string
          description: ONU hardware model
          example: F670LV7.1
        serial_number:
          type: string
          description: ONU serial number
          example: ZTEGC12345678
        rx_power:
          type: string
          description: Received optical power in dBm
          example: "-22.22"
        status:
          type: string
          description: Current ONU status
          enum:
            - Online
            - Offline
            - LOS
            - DyingGasp
          example: Online
      required:
        - board
        - pon
        - onu_id
        - name
        - onu_type
        - serial_number
        - rx_power
        - status
    ONUCustomerInfo:
      type: object
      description: Detailed information for a specific ONU including optical levels and uptime history
      properties:
        board:
          type: integer
          example: 1
        pon:
          type: integer
          example: 1
        onu_id:
          type: integer
          example: 1
        name:
          type: string
          description: Customer or ONU name
          example: Customer-001
        description:
          type: string
          description: ONU location or description
          example: Location
        onu_type:
          type: string
          description: ONU hardware model
          example: F670LV7.1
        serial_number:
          type: string
          description: ONU serial number
          example: ZTEGC12345678
        rx_power:
          type: string
          description: Received optical power in dBm
          example: "-22.22"
        tx_power:
          type: string
          description: Transmitted optical power in dBm
          example: "2.57"
        status:
          type: string
          description: Current ONU status
          enum:
            - Online
            - Offline
            - LOS
            - DyingGasp
          example: Online
        ip_address:
          type: string
          description: ONU management IP address
          example: 10.0.0.1
        last_online:
          type: string
          description: Timestamp of last online event
          example: "2024-08-11 10:09:37"
        last_offline:
          type: string
          description: Timestamp of last offline event
          example: "2024-08-11 10:08:35"
        uptime:
          type: string
          description: Current uptime duration
          example: 5 days 13 hours 10 minutes 50 seconds
        last_down_time_duration:
          type: string
          description: Duration of the last downtime event
          example: 0 days 0 hours 1 minutes 2 seconds
        offline_reason:
          type: string
          description: Reason for the last offline event
          example: PowerOff
        gpon_optical_distance:
          type: string
          description: GPON optical distance in meters
          example: "6701"
      required:
        - board
        - pon
        - onu_id
        - name
        - description
        - onu_type
        - serial_number
        - rx_power
        - tx_power
        - status
        - ip_address
        - last_online
        - last_offline
        - uptime
        - last_down_time_duration
        - offline_reason
        - gpon_optical_distance
    OnuID:
      type: object
      description: An available/empty ONU ID slot
      properties:
        board:
          type: integer
          example: 1
        pon:
          type: integer
          example: 1
        onu_id:
          type: integer
          example: 123
      required:
        - board
        - pon
        - onu_id
    OnuSerialNumber:
      type: object
      description: ONU ID paired with its serial number
      properties:
        board:
          type: integer
          example: 1
        pon:
          type: integer
          example: 1
        onu_id:
          type: integer
          example: 1
        serial_number:
          type: string
          example: ZTEGC12345678
      required:
        - board
        - pon
        - onu_id
        - serial_number
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            status: Bad Request
            error_code: VALIDATION_ERROR
            data:
              message: board_id is not a configured GPON slot
              details:
                received: "99"
            request_id: d7dhp4pciod1cuth4i80
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            status: Unauthorized
            error_code: UNAUTHORIZED
            data: invalid or missing API key
            request_id: d7dhp4pciod1cuth4i80
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 404
            status: Not Found
            error_code: NOT_FOUND
            data:
              message: ONU not found
              details:
                board_id: 1
                pon_id: 8
                onu_id: 42
            request_id: d7dhp4pciod1cuth4i80
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 429
            status: Too Many Requests
            error_code: INTERNAL_ERROR
            data: Rate limit exceeded, try again later
            request_id: d7dhp4pciod1cuth4i80
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 500
            status: Internal Server Error
            error_code: INTERNAL_ERROR
            data: An unexpected error occurred
            request_id: d7dhp4pciod1cuth4i80
...